LPS-63470 move packages
